requirements.txt
absl-py==1.0.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==2.0
flit_core==3.7.1
fonttools==4.32.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
imutils==0.5.4
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
llvmlite==0.38.0
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
natsort==8.1.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==22.0.4
plotly==5.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pynndescent==0.5.6
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==62.1.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
umap-learn==0.5.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
wheel==0.37.1
widgetsnbextension==3.6.0
wrapt==1.14.0
zipp==3.8.0
